GT-1441 Flatten ui to prepare for ConcatAdapter
