--- app/src/main/java/org/cru/godtools/ui/dashboard/tools/ToolsFragmentDataModel.kt
@@ -1,9 +1,0 @@
-package org.cru.godtools.ui.dashboard.tools
-
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-
-class ToolsFragmentDataModel() : ViewModel() {
-    // TODO: Check if has Spotlight Tools
-    val hasSpotlight = MutableLiveData(false)
-}
