+++ app/src/main/res/layout/dashboard_tools_fragment.xml
@@ -23,20 +23,57 @@
                 android:layout_height="match_parent"
                 android:orientation="vertical">
 
-                <androidx.fragment.app.FragmentContainerView
-                    android:id="@+id/fragmentSpotlight"
-                    android:name="org.cru.godtools.ui.dashboard.tools.ToolsSpotlightFragment"
+                <!-- region: Spotlight Section-->
+                <androidx.recyclerview.widget.RecyclerView
+                    android:id="@+id/spotlightRecyclerView"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:orientation="horizontal"
+                    app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
+                    tools:listitem="@layout/dashboard_tools_spotlight_list_item" />
+                <!-- endregion: Spotlight Section-->
+
+                <!-- region: Category Section -->
+                <androidx.appcompat.widget.AppCompatTextView
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
-                    app:visibleIf="@{hasSpotlight}"
-                    tools:layout="@layout/dashboard_tools_spotlight_fragment" />
+                    android:layout_margin="8dp"
+                    android:lineHeight="20sp"
+                    android:text="@string/dashboard_tools_section_categories_label"
+                    android:textSize="22sp" />
 
-                <androidx.fragment.app.FragmentContainerView
-                    android:id="@+id/fragmentCategory"
-                    android:name="org.cru.godtools.ui.dashboard.tools.ToolsCategoryFragment"
+                <androidx.recyclerview.widget.RecyclerView
+                    android:id="@+id/categoryRecyclerView"
+                    android:layout_width="match_parent"
+                    android:layout_height="wrap_content"
+                    android:layout_margin="8dp"
+                    android:minHeight="75dp"
+                    android:nestedScrollingEnabled="false"
+                    android:orientation="vertical"
+                    app:layoutManager="androidx.recyclerview.widget.GridLayoutManager"
+                    app:spanCount="3"
+                    tools:itemCount="5"
+                    tools:listitem="@layout/dashboard_list_item_category" />
+
+                <com.google.android.material.divider.MaterialDivider
+                    android:layout_width="match_parent"
+                    android:layout_height="wrap_content"
+                    android:layout_marginLeft="@dimen/tools_card_margin_horiz"
+                    android:layout_marginRight="@dimen/tools_card_margin_horiz" />
+
+                <androidx.recyclerview.widget.RecyclerView
+                    android:id="@+id/toolsRecyclerView"
                     android:layout_width="match_parent"
                     android:layout_height="match_parent"
-                    tools:layout="@layout/dashboard_tools_category_fragment" />
+                    android:clipToPadding="false"
+                    android:nestedScrollingEnabled="false"
+                    android:orientation="vertical"
+                    android:paddingBottom="12dp"
+                    app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
+                    tools:itemCount="10"
+                    tools:listitem="@layout/dashboard_list_item_tool" />
+                <!-- endregion: Category Section -->
+
             </androidx.appcompat.widget.LinearLayoutCompat>
         </androidx.core.widget.NestedScrollView>
     </androidx.swiperefreshlayout.widget.SwipeRefreshLayout>
